requirements.txt
streamlit==1.24.0
altair<5
boto3==1.26.65
opensearch-py==2.1.1
beautifulsoup4==4.11.2
awscli==1.27.142
plotly
google-search-results
python-dotenv